# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexora Suite API
  version: "1.0.0"
  description: |
    Consolidated OpenAPI specification for the Nexora micro-apps (bookings, routeiq, home, etc.).
servers:
  - url: http://localhost:5000
    description: Local development server (adjust as needed)
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created
        '400':
          description: Bad request
  /api/auth/login:
    post:
      summary: Login and receive access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
...
components:
  schemas:
...
    UserRegister:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    UserLogin:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...